requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-redis==5.0.0
djangorestframework==3.12.4
environs==9.3.3
marshmallow==3.13.0
mysqlclient==2.0.3
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
